One PatchFileChangeChecker, so version conflict warnings only appear once
